remote/requirements.txt
accelerate==1.0.1
annotated-types==0.7.0
blinker==1.8.2
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
filelock @ file:///croot/filelock_1700591183607/work
Flask==3.0.3
fonttools==4.58.0
fsspec==2024.10.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
huggingface-hub==0.26.1
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.5.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.4
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1717597493534/work
numba==0.60.0
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp39-cp39-linux_x86_64.whl#sha256=d86a49760b169e0c4fb8c00d248077a0474f640071b7ef584afd5ad4f03b9428
packaging==24.1
pandas==2.2.3
peft==0.14.0
pillow @ file:///croot/pillow_1721059439630/work
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pynndescent==0.5.13
pyparsing==3.2.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.2.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
transformers==4.46.0
triton==3.1.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2025.2
umap-learn==0.5.7
urllib3 @ file:///croot/urllib3_1727769808118/work
Werkzeug==3.0.6
zipp==3.20.2
